factor out logic for handleClipDragEnd
